move the permissions check to new file
