+++ facia-tool/app/controllers/FrontController.scala
@@ -62,22 +62,5 @@ object FrontController extends Controller with PanDomainAuthActions {
   }
 }
 
-import play.api.mvc._
-
-object PermissionCheckAction extends ActionFilter[UserRequest] {
-  override def filter[A](request: UserRequest[A]) = {
-    import scala.concurrent.ExecutionContext.Implicits.global
-    for {
-      b <- PermissionsReader.get(SimplePermission.ConfigureFronts, request.user)
-    } yield
-    (if(b) {
-      None
-    }
-    else {
-      Logger.info("user not authorized to configure fronts")
-      Some(Results.Unauthorized(views.html.unauthorized()))
-    })
-  }
-}
